requirements.txt
Flask==3.1.3
flask-ldap3-login==1.0.2
Flask-Login==0.6.3
psycopg2-binary==2.9.11
uWSGI==2.0.31
requests==2.32.5
